requirements.txt
# ==== LLM & Google Generative AI ====
google-generativeai
google-cloud-aiplatform

# ==== Embeddings (Optional) ====
sentence-transformers
transformers
torch

# ==== Vector Store ====
chromadb

# ==== Text Processing & LLM Orchestration ====
tiktoken
unstructured
langchain

# ==== Backend (API Server) ====
fastapi[all]
uvicorn

# ==== Frontend (Optional UI) ====
streamlit

# ==== Document Parsing (Optional) ====
pdfplumber
beautifulsoup4

# ==== Data Utilities ====
pandas
numpy

# ==== Environment & Config Management ====
python-dotenv

# ==== JSON & Typing ====
orjson
typing_extensions
